requirements.txt
discord.py==1.7.3
python-dotenv==0.17.0
loguru==0.5.3
pymorphy2==0.9.1
pymorphy2-dicts==2.4.393442.3710985
DAWG-Python==0.7.2